[captures2024-captures-2024#79] refactor: dto 패키지 구조 리팩토링
