+++ apps/soongan-api/src/main/kotlin/com/soongan/soonganbackend/soonganapi/interfaces/auth/dto/response/LoginResponseDto.kt
@@ -1,4 +1,4 @@
-package com.soongan.soonganbackend.soonganapi.interfaces.auth.response
+package com.soongan.soonganbackend.soonganapi.interfaces.auth.dto.response
 
 data class LoginResponseDto(
     val accessToken: String,
