+++ apps/soongan-api/src/main/kotlin/com/soongan/soonganbackend/soonganapi/interfaces/postLike/dto/request/PostLikeRequestDto.kt
@@ -1,4 +1,4 @@
-package com.soongan.soonganbackend.soonganapi.interfaces.postLike.dto
+package com.soongan.soonganbackend.soonganapi.interfaces.postLike.dto.request
 
 import com.soongan.soonganbackend.soonganpersistence.storage.comment.ContestTypeEnum
