+++ apps/soongan-api/src/main/kotlin/com/soongan/soonganbackend/soonganapi/interfaces/postLike/dto/response/PostLikeResponseDto.kt
@@ -1,4 +1,4 @@
-package com.soongan.soonganbackend.soonganapi.interfaces.postLike.dto
+package com.soongan.soonganbackend.soonganapi.interfaces.postLike.dto.response
 
 data class PostLikeResponseDto(
     val postId: Long,
